% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixEffects.r
\name{summary.l2test}
\alias{summary.l2test}
\title{Summary Method for L2 Test Results}
\usage{
\method{summary}{l2test}(object, ...)
}
\arguments{
\item{object}{An l2test object}

\item{...}{Additional arguments}
}
\description{
Summary Method for L2 Test Results
}
